% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrvfunction.R
\name{rrv}
\alias{rrv}
\title{computes reliability, risk, resilience, vulnerability from drought index dataset}
\usage{
rrv(X, cutoff)
}
\arguments{
\item{X}{a vector of sequential drought index values (numeric)}

\item{cutoff}{numeric value indicating the drought cutoff for the index}
}
\value{
rrvx reliability, risk, resilience, vulnerability
}
\description{
computes reliability, risk, resilience, vulnerability from drought index dataset
}
\examples{
rrv(rnorm(1000),cutoff=-0.7)
}
\references{
Hashimoto, T., Stedinger, J.R. and Loucks, D.P., 1982. Reliability, resiliency, and vulnerability criteria for water resource system performance evaluation. Water resources research, 18(1), pp.14-20. 
\url{http://pure.iiasa.ac.at/id/eprint/14077/1/wrcr3066.pdf}\url{http://pure.iiasa.ac.at/id/eprint/14077/1/wrcr3066.pdf}
}
\author{
Venki Uddameri, \email{venki.uddameri@ttu.edu}
}
